+++ src/java/org/malbino/orion/facades/NotaFacade.java
@@ -125,7 +125,7 @@ public class NotaFacade extends AbstractFacade<Nota> {
 
         return l;
     }
-    
+
     public List<Nota> listaNotasPruebaRecuperacion(Inscrito inscrito) {
         List<Nota> l = new ArrayList();
 
@@ -370,6 +370,23 @@ public class NotaFacade extends AbstractFacade<Nota> {
         return l;
     }
 
+    public List<Nota> listaNotasFaltantesPRAE(int id_gestionacademica, int id_carrera) {
+        List<Nota> l = new ArrayList();
+        try {
+            Query q = em.createQuery("SELECT n FROM Nota n JOIN n.gestionAcademica ga JOIN n.materia m JOIN m.carrera c JOIN n.estudiante e WHERE ga.id_gestionacademica=:id_gestionacademica AND c.id_carrera=:id_carrera AND m.curricular=:curricular AND n.modalidad=:modalidad AND n.nf IS NULL ORDER BY e.primerApellido, e.segundoApellido, e.nombre");
+            q.setParameter("id_gestionacademica", id_gestionacademica);
+            q.setParameter("id_carrera", id_carrera);
+            q.setParameter("curricular", Boolean.TRUE);
+            q.setParameter("modalidad", Modalidad.PRUEBA_RECUPERATORIA_ADICIONAL_EXCEPCIONAL);
+
+            l = q.getResultList();
+        } catch (Exception e) {
+
+        }
+
+        return l;
+    }
+
     public List<Nota> listaRegistroNotasSemestral(int id_gestionacademica, int id_carrera, TipoNota tipoNota) {
         List<Nota> l = new ArrayList();
         try {
@@ -423,7 +440,7 @@ public class NotaFacade extends AbstractFacade<Nota> {
 
         return l;
     }
-    
+
     public List<Nota> listaRegistroNotasRecuperatorio(Inscrito inscrito) {
         List<Nota> l = new ArrayList();