+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.5.0] - 2023-07-06
+### Added
+- Centralizador Calificaciones PRAE
+### Changed
+- Centralizador Calificaciones
+### Removed
+### Fixed
+
 ## [2.4.1] - 2023-06-21
 ### Added
 ### Changed